// src/types/scalp.ts
export type ScalpMbtiType =
  | 'oily_trouble_type' // 트러블 폭풍형
  | 'oily_sensitive_type' // 지성 민감형
  | 'oily_scaling_type' // 지성 비듬형
  | 'seborrheic_type' // 깐깐 지성형
  | 'dry_trouble_type' // 건조 트러블형
  | 'dry_sensitive_type' // 민감 건조형
  | 'dry_scaling_type' // 건조 비듬형
  | 'balanced_type' // 밸런스형